chore(context): move api to appropriate django app
The API belongs to the same Django App that has its used context
processor.
