+++ backend/core/views.py
@@ -1,8 +1,9 @@
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
+from .context_processors import domain_name, site_owner, social_media_links
 from .models import About
-from .serializers import AboutSerializer
+from .serializers import AboutSerializer, SocialMediaLinkSerializer
 
 
 class AboutView(APIView):
@@ -12,3 +13,20 @@ class AboutView(APIView):
         except About.DoesNotExist:
             return Response({})
         return Response(AboutSerializer(about).data)
+
+
+class FrontendContextView(APIView):
+    """API for providing previously used template context processor injections."""
+
+    def get(self, request):
+        social_media_queryset = social_media_links(request)["social_media_links"]
+
+        return Response(
+            {
+                "domain_name": domain_name(request)["domain_name"],
+                "site_owner": site_owner(request)["site_owner"],
+                "social_media_links": SocialMediaLinkSerializer(
+                    social_media_queryset, many=True
+                ).data,
+            }
+        )
